backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
openai==1.37.0 # optional for LLM generation